+++ team/src/main/java/sansam/team/test/command/dto/TestDTO.java
@@ -1,4 +1,4 @@
-package sansam.team.test.query.servicies;
+package sansam.team.test.command.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor  // 기본 생성자
 @AllArgsConstructor // 모든 필드를 받는 생성자
-public class TestQueryDTO {
+public class TestDTO {
 
     private String content;
 }
