+++ team/src/main/java/sansam/team/test/command/repository/TestRepository.java
@@ -1,7 +1,7 @@
-package sansam.team.test.command.repositories;
+package sansam.team.test.command.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sansam.team.test.command.entities.Test;
+import sansam.team.test.command.entity.Test;
 
 public interface TestRepository extends JpaRepository<Test, Long> {
 }
