+++ team/src/main/java/sansam/team/test/query/service/TestQueryService.java
@@ -1,9 +1,9 @@
-package sansam.team.test.query.servicies;
+package sansam.team.test.query.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import sansam.team.test.command.services.TestDTO;
-import sansam.team.test.query.mappers.TestMapper;
+import sansam.team.test.command.dto.TestDTO;
+import sansam.team.test.query.mapper.TestMapper;
 
 import java.util.List;
